use named exports for library modules
