--- jest.config.js
@@ -1,12 +1,0 @@
-module.exports = {
-  preset: "ts-jest",
-  // transform: { '^.+\\.ts?$': 'ts-jest' },
-  transform: {
-    "^.+\\.(ts|tsx)$": "ts-jest",
-    "^.+\\.(js)$": "babel-jest",
-  },
-  transformIgnorePatterns: [],
-  testEnvironment: "node",
-  testRegex: ".*\\.(test|spec)?\\.(ts|tsx)$",
-  moduleFileExtensions: ["ts", "tsx", "js", "jsx", "json", "node"],
-};
