+++ src/__tests__/axios.test.ts
@@ -1,6 +1,6 @@
 import { Payload } from "../payload";
-import { APIKEY, EmptyClientMetadata } from "./index.test";
-import APIToolkit, { ReportError } from "../index";
+import { APIKEY, EmptyClientMetadata } from "./apitoolkit";
+import APIToolkit, { ReportError } from "../apitoolkit";
 import request from "supertest";
 import express, { Request, Response } from "express";
 import axios, { AxiosResponse } from "axios";
